requirements.txt
pandas
datasets
transformers
huggingface
evaluate
torch
numpy
rouge-score
sentencepiece
accelerate
bitsandbytes
tqdm
peft
trl
scikit-learn
pyyaml
matplotlib
deepspeed